rename type.js in store to actions.js
